Fix tests and probable regression
do loop, but only create one new connection (if we created one and
it didn't work, creating another is unlikley to work better). We
need to loop though otherwise we'll fail unnecessarily if we just
had a retired connection (since we rely on the
ConnectionDeadException to remove it from use).